src/requirements.txt
Beaker==1.6.4
PyYAML==3.10
argparse==1.2.1
bottle==0.11.5
wsgiref==0.1.2
